Update 2.1.1.4.
My software is now obsolete, time to switch to Flames! ^w^

+++ Sparks/Config/OtherAttributes.cs
@@ -69,7 +69,7 @@ namespace GoldenSparks.Config {
         
         public override string Serialise(object value) {
             LevelPermission perm = (LevelPermission)value;
-            return ((sbyte)perm).ToString();
+            return ((int)perm).ToString();
         }
     }
